add .env in gitignore

+++ services/order/src/main/java/com/rishav/ecommerce/orderline/OrderLineRepository.java
@@ -1,9 +1,8 @@
 package com.rishav.ecommerce.orderline;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-
 import java.util.List;
 
-public interface OrderLineRepository extends JpaRepository<OrderLine, Integer> {
-    List<OrderLine> findAllByOrderId(Integer orderId);
+public interface OrderLineRepository extends JpaRepository<OrderLine, Long> {
+    List<OrderLine> findAllByOrderId(Long orderId);
 }
