+++ services/order/src/main/java/com/rishav/ecommerce/payment/PaymentRequest.java
@@ -9,7 +9,7 @@ import java.math.BigDecimal;
 public record PaymentRequest(
         BigDecimal amount,
         PaymentMethod paymentMethod,
-        Integer orderId,
+        Long orderId,
         String orderReference,
         CustomerResponse customer
 ) {
